Create detect_anomalies.py
The following script was a function I  developed to detect anomalies on any numerical column of data passed as a parameter. The user can also specify the threshhold to which they want results. Finally, result can be plotted using 2 px.line traces, or as the user pleases.

In the above, I passed a dataframe that had a column of foot traffic counts at Giant stores around the DC area. The dataframe was an aggregate traffic index of all people at shopping at giant stores between 2019 - Present