Create get_ports.py
This function is a script that can be used to extract information about ports and ships that are currently moored. 

It starts by importing several libraries, including the pandas library which is used for data manipulation and analysis. 

It then takes four parameters as input - project_name, pings, tracks, and call_sign. All these inputs are expected to be CSV files. The function then reads the contents of the CSV files into four pandas dataframes: geolocation_pings, emr_tracks, and call_sign. 

It then renames the column 'device_id' in geolocation_pings to 'mmsi' for consistency and then filters the required columns in geolocation_pings .

 It then finds the unique rows from geolocation_pings dataframe. 

The next step is to merge two dataframe geolocation_pings and emr_tracks on the column 'mmsi' with how = 'right' and the new dataframe is named results. 

Then it updates the column 'series' and renames the column 'Allocated to' to 'Country' in call_sign dataframe. 

It then creates a new column 'callsign_new' in results dataframe with the updated column 'callsign' and then extracts the first two elements of the column 'callsign_new' and stores them in the same column. 

The column 'callsign_new' is then renamed to 'Series' and the results dataframe is then merged with call_sign dataframe on column 'Series'. 

Then it filters the dataframe final for the rows where the nav_status is "Moored" and sog is zero and the dataframe ports is stored. 

Finally, ports dataframe is saved as CSV file with the given name project_name + '.csv' in the output_folder and the message 'Completed, output saved' is printed.