yes
It's possible that using Spark could speed up the process of pulling data from the API, but it depends on the size of the data and the resources available for processing. Spark is designed for distributed computing, so if you have a large dataset and access to a cluster of machines, it could potentially speed up the processing time. However, if the dataset is relatively small and you're running on a single machine, the overhead of setting up and running Spark might actually slow down the process.
If you do decide to use Spark, you'll need to convert your code to work with Spark's data structures, such as DataFrames or RDDs. You'll also need to set up a Spark cluster and configure it appropriately for your use case. This can be a non-trivial task, so it's important to carefully consider whether the potential speedup is worth the added complexity.
Another alternative to consider is using asynchronous requests to the API instead of making synchronous requests. This would allow you to send multiple requests in parallel and speed up the data retrieval process. You could use the asyncio library in Python to implement this approach.